Move processSwapQuoteError from actions file to scene file
